requirements.txt
django==4.1.6
djangorestframework==3.14.0
pika==1.3.1
scikit-learn==1.2.1
matplotlib==3.7.0
seaborn==0.12.2
requests==2.28.2